refactor(oid4vc): split test_improvements into test_dpop, test_did_configuration, test_pex_jwt_vp, test_metadata
